requirements.txt
mysql-connector-python==8.0.33
openai==0.27.8
python-dotenv==1.0.0
